Launch the instant run service before launching activity
If the activity is launched first, and stops at a breakpoint, then the
subsequent launch of the service could raise an ANR. This is a short
term fix where we launch the service first. With this fix, the issue
could still arise if users set a breakpoint in Application.onCreate()
for instance. A better fix will be to remove this service and launch the
instant run server directly from a ContentProvider, and will be
performed in a subsequent CL.

Bug: https://code.google.com/p/android/issues/detail?id=231497
Test: Manual, set a breakpoint in onCreate(), launch app under debugger,
and just wait there for > 5s.

Change-Id: Ica71f4f13533606d10de28771f3737e0bd0f23d5
(cherry-picked from ae921dfbd and fixed conflicts in
android/src/com/android/tools/idea/run/AndroidLaunchTasksProvider.java)