Handle the case gradle build cancelled in GoToApkLocationTask
Fix b.android.com/229315
The problem was buildError instance was only set for exception type
of BuildException, not set for other Throwable types.
For example, when BuildCancelledException, which is not BuildException
type, happened, buildError is null, thus GradleInvocationResult was
considered successful.

The fix added a new variable GradleInvocationResult#myBuildCancelled. If
build was cancelled, then bubble shows notification "Build cancelled".

Test: GoToApkLocationTaskTest.java
Change-Id: Iff2b73d17ed9be30736efdcb184959e43c1b862d