Add navigation capabilities to class and instance view
This CL adds IntelliJ navigation capabilities to the memory
profiler's class and instance views. This allows users to
right click on a class/instance and navigate to the implementation
of the selected instance.

Furthermore, the stack view has been unified between memory and
network profilers, resulting in a unified component that can
handle navigation on callstacks.

There is also some minor work to un-maximize the tool window when
navigation occurs.

Test: Added tests for navigation in addition to other test fixes.

Change-Id: I851456839df916c69d1f1cfc377ef5b1f0f09765