Parse capture in a different thread
Parsing the CPU capture shouldn't block the UI thread. However, we need
to make sure to update the UI once a capture is done.

This CL also fix the crash that was happening when a short capture was
taking. Instead of crashing, now it's logging the error and closing the
capture view panel.

Also, a minor bug with the Stop/Record button was fixed. The button
label was being set to Stop once hitting Record and was never being
set back to Record.

Bug: 33431119
Test: CpuProfilerStageTest. CpuProfilerStageViewTest is on its way
Change-Id: I5dad9fd5b33215eae5f77d0d3492284b0cafa444