Fix typing in EditorTextField
There is a small period of time, after the EditorTextField is created
and displayed, that key typing can be lost. EditorTextField is a very
complex editor (basically a full fledged Editor that is squeezed in a
single line) and there is no good way to find out when is ready to
type.

Test: Run RefactoringFlowTest@testResourceConflict(). Was failing 95%
in windows (100% after the first run)
Change-Id: I0a7a60781b500fa2a454d234160784ce3032fbb7