Fix AndroidToolWindowFixture.selectItem
Current ProcessListFixture.selectItem fails to work properly as the
items in drop-down combox could vary extremely fast during the selection.
So, the identified index in EDT could be invalid at the fixture's selection
point.

Fix it by programatically selecting the item in the same EDT to query
and identify the item.

Test: affect several tests and thus CL should not introduce new issue in
PSQ.
Change-Id: I0e439a08e2fe194d248db0c46d5a6f3351ada696