Migrated legacy allocation tracking to perfd-proxy layer
Instead of shoehorning the legacy tracking workflow into perfd-host,
the logic now lives in a proxy layer between perfd-host and perfd.

Perfd-host is now oblivious to the legacy workflow, and the updated logic
to gather the allocation info and data is consistent with heap dumps.

The APIs for retrieving allocaiton events, classes and call stacks have
also been refactored to be more easily portable to a streaming mode.

Test: added+updated unit tests for MemoryServiceProxy + MemoryDataPoller
Additional tests to follow on MemoryTable

Change-Id: Id020bd34ade6215aab93083671246746379a7ae3