UI Tests: More resilient typing
The default FEST typing (which uses the Java Robot) is now broken in
the latest Mac OS update (Sierra).
Sometimes, tests fail because characters "disappear" during typing.
To make tests more resilient to typing, this CL replaces some of the
typing with direct Java calls to setText(), and others (where setting
text programmatically is not possible) with the FEST typing version
that sends typing events directly to Java (instead of going through
the Operating System)

Test: Mac, any UI test that types, eg AddGradleDependencyTest
Change-Id: I625d59c32c85f3b37d5828bd090ecef1c4a7cf6b