Add design lib dependency when using AppBar dialog
The AppBar config dialog displays a preview. Since the adding of the
dependency was moved to after the dialog exits, the preview was always
broken for this component unless the user manually added the design
dependency.
This CL makes the dialog to add the dependency if missing and updates
the preview once the compilation has finished.

Bug: http://b.android.com/214653
Test: Added UI test for the AppBarConfigurationDialog
Change-Id: I980c1cbb1dd1b4f90399f4a35a6e3c6b65ed9a4f