Log UNKNOWN thread states
Right now we're crashing the application when we get an unexpected
thread state from perfd. The state chart can't map it to a UI state and
throws an exception. A simple log containing the state that caused it to
crash might help us in the future if the exception is thrown again.

Test: no need for new tests. Just adding a Log.
Change-Id: Iab394a23e9ee1788795a8eb95b4752b39690908d