Removed unnecessary setup steps when using model from disk cache.
When reopening a project, the IDE will set up the project
using the Gradle models obtained from the disk cache. Since
we are using the cached models, only setting the models
in facets and validating state of the project make sense.
This change improves sync performance when reopening a
project.

Test: existing tests should pass.
Change-Id: Ic09aae3fe7027d09692a03ddbcaad9f50ff7e87b