Heap's Reference/Field objects to inherit from HeapDumpInstanceObject
This CL is a prequel to being able to handle selection of a field/reference
objects and put them as the current selected class + instance. By inheriting
from HeapDumpInstanceObject, we get a lot of the Class/Instance information
for free from the base class implementation, so we don't need to do further
processing in the Reference/Field classes.

Test: unit test

Change-Id: Id6f0ca5940f14b9c2aac4b9970f9f79b4ead0cbd