Allow SpeedSearch to function across groups in the palette.
Before this change SpeedSearch would only find component names
inside a single group in the palette.
The fix is a refactoring of TreeGrid such that the listeners of
TreeGrid are being notified when any of individual JList
components causes: KeyEvents, MouseEvents, ListSelectionEvents,
FocusEvents.
That allows the base implementation of SpeedSearch to function
with a TreeGrid instance.

Bug: http://b.android.com/232282
Test: Updated TreeGridTest and added a UI test to test the
      speed search functionality.

Change-Id: I4a0e6338e77cbaa9a090d4e812fdf20ffcc963aa