Fix the property editor tests and reenable them.
We were emulating typing in a text box where focus is required.
There is code in JTextComponent where a focus gain causes the
component instance to be registered with AppContext with a
private key. The only way to unregister the instance is to call
removeNotify. Without it we would leak project instances with
reference from AppContext. This has been fixed.

The handling of focus caused problems when NlEnumEditorTest
was being run before NlReferenceEditorTest. The emulation of
focus has been changed to mimic the swing UI closer.

Test: NA
Change-Id: Ib99feddd99e3474ee51bc98c8662b3c1bdbc77c9