Delay the creation of NlModel until the AndroidModel exists.
For Gradle projects: delay the creation of NlModel until
AndroidModel is created. The logic is moved from
NlModel.requestModelUpdate to NlEditorPanel.

By doing this we avoid loading Aar libraries from the
Intellij dependency files but instead use gradle to locate
the dependencies for loading resources.
An assert for this has been added.

The delay of the NlModel creation will cause a loading
animation to show up while the project is build/indexed.

Test: No unit tests added. The loading animation can be
      manually reproduced by changing the build.gradle
      file manually and then close/reload the project
      without first rebuilding.

Change-Id: If7c79886417fd3b1c0fa449cb03f3593f92ff55b