Move the srcCompat logic to a more general place.
The decision of whether to use src or srcCompat is needed both
in the ImageViewHandler and in the property system.
This CL consolidates the logic for this decision in the
ImageViewHandler.

The new tests in NlPropertiesTest needed to override the
ViewHandlerManager component in the ComponentManager which is
different from overriding a component in the PicoContainer.
Updated PicoComponentStack to handle this case.

Bug: https://buganizer.corp.google.com/issues/32817336

Test: Unit tests have been updated.

Change-Id: I8d1bbf94aa0eb0753b6c7e12d5ba59d997e66933