New ValidatorPanel method for registering observable messages
A common pattern is for a UI class to create a String message which is
the target for some validation string. If the String is set, the UI
state should be considered invalid.

A user could currently work around this by doing:

validator.register(message, message -> if (message.isEmpty()) ...)

but this seems like something people might do enough it's worth adding a
shortcut for it.

Test: N/A, UI change covered by existing tests.
Change-Id: If794feda5fce9072222f720ee459defbfb42c3b7