Show all the resources across all the folders
The translations editor showed a single row for multiple resources with
the same name in multiple folders. Change the resource key from the name
to the name and resource folder (StringResourceKey). Manually manage the
ModuleResourceRepository children in StringResourceRepository because
LocalResourceRepository.doMerge merges resources with the same name.

The translations editor will create new translations in the resource's
folder instead of always in the primary one.

Bug: http://b.android.com/75903

Test: TranslationsEditorTest.enteringValueUpdatesDebugStringsXml and
      enteringValueUpdatesMainStringsXml

Change-Id: I06f5eb4f3441215593f5c99e5909a176943977e9