Remove read lock from RenderTask calls
Holding the lock for long periods of time causes problem in the rest of
the IDE. We "shouldn't" need to hold the lock at this level as the lock
should be acquired closer to the actual use (probably within
LayoutlibCallbackImpl).
Also, added a couple of assertions to check if we still need the ugly
retry looks.

Test: Covered by existing tests
Change-Id: I7b1bb369016bd5944095d50cd1bd196e7c74aee7