Changing default screen density for tests
In layout tests right now we have two different Configurations:
the real one in NlModel, and the mock created in
LayoutTestUtilities.createScreen(). Unfortunately, by default these have
different Density values: XHIGH for the one in NlModel, and MEDIUM for
the mock.
We should only use one Configuration. To prepare for that change,
switching the density used in the mock to XHIGH (which is better than
MEDIUM for testing since in MEDIUM 1px=1dp, and so that logic isn't
excercised).

Test: this is a test-only change
Change-Id: If18502f292ce383f2dae7830bf8e5d10c70fe63f