Fixing deadlock when using SyncNlModel
We were running into a 10-second deadlock when using SyncNlModel and
doing drag-and-drop actions, since DragDropInteraction.moveTo() runs
myDragHandler.commit in a write action, which can eventually kick off
layoutlib rendering in a separate thread, which then tries to get a read
lock for a while.
Changed so that the normally-async requests are kicked off after the
action is done rather than trying to do it within the action itself.
This saves many minutes, for example AbsoluteLayoutHandlerTest went from
183 seconds on the most recent postsubmit to 2 seconds for this change's
presubmit.

Test: this is a test-only change
Change-Id: Id5652709a55a549f38aa3f6ba16b1fa2303f94b7