Fixed timing issue related to heap dump and allocation tracking
Currently we depend on the HeapDumpInfo to already exist when we query
for the dump data. But since the data poller runs on a fixed timer, we
might not get the HeapDumpInfo back in time before the UI starts asking
for the dump.

The fix is to insert the HeapDumpInfo immediately when requesting
heap dumps if the request was successful.

The same scenario also applies to allocation tracking.

Bug: 34199717
Test: added unit test
Change-Id: I664bb11fa922b2f76729203c0e0de19debff5b0a