Wait until Gradle build is done before showing demo images.
AppBarConfigurationDialog:
We used to paint the demo images in when GradleSync reported back
that it was done. It turns out this callback is activated when
the first Gradle invocation has finished. LayoutLib was unable
to produce a demo layout without errors. Now wait until the Gradle
build is done.
Display a spinner while waiting.

Also: Place the FAB at the bottom right instead of on the AppBar.

Test: This should make the AppBarConfigurationDialogTest more
      predictable.

Change-Id: I2279c6619e6559c49d470b4387cc24ac02630ed9