Supporting Plug/Unplug scenarios.
This change adds a timestamp that gets set on the process proto,
this timestamp is to indicate when a process is created.
The timestamp is then used to build the timeline for that process.
In this iteration the process timestamp is reset on each connect/disconnect
and a todo is in place of getting the process start time.

Test: ProfilerServiceTest, Profilers Test
Change-Id: If899e8954a78df0c55a131a1d7484c32a49569ca