Force CompilerOutputModuleSetupStep when reopening project
ModuleSetupStep#invokeOnSkippedSync was introduced in http://ag/1797567
in order to skip unnecessary setup steps when a project is reloaded.

Skipping CompilerOutputModuleSetupStep caused error in running Unit
tests since CompilerSettingsSetup#setOutputPaths was not called and then
intermediate classes location was not in the classpath when running
tests and then no tests were found, resulting in an empty test suite.

This CL fixes this problem by invoking CompilerOutputModuleSetupStep
even when Sync is skipped (when reloading a project).

Test: AndroidJUnitClasspathTest and CompilerOutputModuleSetupStepTest

Fixes bug http://b.android.com/233410

Change-Id: I48bd3f673fb487374beae4506796ecf77da59810