Use start time as id for heap dump and allocation tracking
Previously we used a incremental int counter on perfd as the dump
id which gets reset everytime perfd restarts. This results in data
collision. To ensure the id is unique in each session/process combo,
we now use the StartTime of the heap dump/allocation tracking request
as the id for HeapDumpInfo/AllocationsInfo respectively.

Bug: 34892680
Test: updated test.
Change-Id: Ie2d712948654d820b657cb67e02bd2046c9922f8