Correctly set up gradle and the plugin in TemplateTest
This updates the plugin and gradle wrapper to the latest
version before a build and then performs a sync and imports
the project. Without a sync / import Gradle is not set up
with the correct paths to find the plugin@HEAD that we
use for testing and so the project will not build.

This mirrors the way we test projects in other parts of
the source base and tests the Templates against the
version of the plugin / gradle that they will be released
against.

Fixes: http://b.android.com/240097

Test: N/A
Change-Id: I7dcd77228f9cc0445b2c25c8ff3dcd1f89d24cfc