Implement Dump DisplayList in Inspector
Making the dump displaylist request requires having a reference to the
client which layout inspector didn't have.
Updated LayoutInspectorAction to pass along the client/window to
LayoutInspectorContext once the .li file is saved.
So the option is enabled when user opens layout inspector from a device
but disabled if layout inspector is opened from a saved .li file.

Feature is hidden behind a env flag (-Ddump.displaylist.enabled=true) until we get UX/PM input on how to
display the result to users.

Fixes: http://b.android.com/223546
Test: Manually tested
Change-Id: I410163e6c59b866c4c0548cee6c6f9f6206b0fe7