Format templates only after project is initialized
After Ifb05e407da14aaf283633d2b5ffdc2046a9ce86a we saw an issue with c++
formatting for new projects.  It was caused by templating running before
symbols are available, causing spacing logic to be broken.

Make sure not to render templates until the project is initialized.
Ideally we would not have this order dependency here, but breaking it
will be complex.

Test: Ran NewModuleTest, added new test there.
Change-Id: I5146481770dcb78dd795f3ee101710e380194215