First part of scene creation refactoring
This moves the existing scene-creation and -updating related logic out
of Scene and into a builder.
In a subsequent change, similar logic will be moved out of NlModel into
the builder.
The goal is to have both Scene and NlModel be non-layout specific (and
clean things up slightly in the process).

Test: No new logic, just moving things around. Existing tests updated
accordingly.

Change-Id: Id86c101f7a4c3c40de8984e1b90b1c8f705072ac