Now handling disconnected devices and killed processes.
It can be very useful for a user to inspect a process after it crashes.
Prior to this CL, once the process died, it was immediately discarded
from the process pulldown. Now, we leave the process in, marking it
[DEAD].

If the user restarts the app or disconnects the device, then the new
process will take focus and the [DEAD] one will be discarded, as expected.

Bug: 33430382
Test: StudioProfilersTest added to cover new code

Change-Id: Ibaf65f1beb71d2d0cecb325049623ef99f7f57f0