Fix blurry layoutlib render image on mac
It turns out that we still need to do a specific scale for retina
displays so I reused the old method to get the Retina scaled image.

This does not affect the fix for the out of memory error when zooming.

Test: re-run tests for ScreenView Layer and manually tested the quality
of the image on a retina display and on a normal display. Also verified
that we can zoom up to 2000% on mac with no out of memory issue.

Change-Id: I530400cd10a699ae5f9a463d3a3c267b158ea319