Reactivate sharding for android-tests and designer tests.
Bazel post submit frequently times out so we are looking into ways to
speed up our test. Taking off the exclusive tag from android-tests and
designer-tests should help because it alows android-tests which is one
of the most time consuming targets to run along with other targets.
Sharding it should make things even faster.

I got the shard numbers from what they were before we got rid of them.
I feel they are a bit conservative for 840s.

In the mean time we lose a little coverage because we need to disable
the tests that fail due to non-hermeticism:

There are 3 android-tests tests and 1 designer test failures.
This CL disables them.

Test: Bazel test android-tests and designer-tests.
Change-Id: I236138b4339400402f10bbfd73b39415f5559bb6