[APK Debugging] Implemented navigation from dex class to source java file.
With this CL, when the user double-clicks on a Java node
in class.dex node (in the "Android" view), and if the
user has specified the paths for Java source code, the
IDE will open the corresponding .java file in the editor.

For example, if the user opens the 'MyActivity' node, the
IDE will open 'MyActivity.java' file in the editor, if the
file is found, and if both package and class names match.

Test: JUnit test.
Change-Id: Ic7eaa68c7c62048d851273bb1a3d9a20fc1f68a9