Add NPW UI for new Instant Apps project structure
This CL updates the New Project Wizard and New Module
Wizards to enable them to work with the new templates
that output the module structure described in
http://b/34667965

This CL adds support for Instant Apps to both the old
and new flows of the NPW (although the new flow is not
yet fully functional)

It also tidies up some of the naming (I am moving from
AIA to InstantApp throughout) and starts normalising
InstantApps utility code in to the InstantApps helper
class, particularly with an eye to those operations that
are temporary work-arounds for future work such as the
SDK Manager integration.

Test: NewProjectTest and InstantAppTest in the GUI tests
exercise these code paths and have been updated
appropriately. Utility code has had tests added to
InstantAppsTest

Change-Id: I334381560e5848bbbfee85f2c5cce48f88bc82d7