[APK Debugging] APK file is no longer root node in "Android" view.
With this CL, the root node of the "Android" view represents
the root module, but not the APK file. Theh problem was that
if the node representing the APK has children (it had when it
was the root node), double-clicking the node won't open the
APK file in the editor, but it will just show its children
nodes.

Test: JUnit tests (in android-apk module.)
Change-Id: I687209ab7a6b030a68638716baccbf69da7fa543