Cache CPU-related responses to reduce database queries
Sequential queries to cpu service often pass the same arguments to the
request (start/end timestamp, process ids, etc). For instance, getData
is called in sequence with the same arguments to obtain app and system
usage in the CPU usage chart. Also, the thread monitor call getThreads
and getTraceInfo with the same request arguments for each thread.
Caching the response in these scenarios drops the response time of the
service considerably.

Some data obtained after running DisplayingBitmaps for 15 to 25 minutes
with moderate activity is presented in the table below. Four samples of
each method were collected (2 cached and 2 uncached).

 Method       | uncached  | cached  | x Faster
 -------------+-----------+---------+----------
 getData      | 677.481us | 8.799us | 75+
 getData      | 781.357us | 8.859us |
 -------------+-----------+---------+----------
 getThreads   | 1.140ms   | 9.552us | 100+
 getThreads   | 1.532ms   | 7.786us |
 -------------+-----------+---------+----------
 getTraceInfo | 67.839us  | 7.320us | 8+
 getTraceInfo | 60.754us  | 7.584us |

Test: Unit tests added to CpuDataPollerTest
Bug: 35435729
Change-Id: I1d0fe85fcd1af93a06c1980ec9a6610031eeb1d0