Fix UI tests after merge with IntelliJ 2017
JavaDirectoryServiceImpl@createClassFromTemplate is now asserting
that it's not running on under a write lock.
This fix changes CreateFileFromTemplateDialog to use a
ElementCreator that doesn't run on a write lock.
Though we no longer need a write lock to create a file from a template
we still need the lock to call createPackageSubdirectory

Test: CreateNewClassDialogGuiTest
Change-Id: Ica2f3a30a67b506810935b454671621333a01949