Fix focus problem in the NlPreviewForm
When performing an activation of the NlPreviewForm, after the palette
has been opened and closed, the focus cannot be properly given to the
preview. This is fixed by forgetting what component in the preview had
the focus when last activated, and simply giving it each time to the
main component.

Test: NlPreviewTest
Change-Id: I90c60ee9d2efdc4689bd88eb195c81d99815c00e