Updating Datastore Session column to be an integer for performance.
Noticed a 30ms gain in performance when moving from String column to
Integer column for Session information. The speed slow down was only
encountered on my machine, and I only noticed after I started setting
the Boot Id to the UUID that is pulled from the device. To be commited
in a follow up CL.

Test: Perfd-host test.
Change-Id: Idaf36e94d84dd4b4eaf5159788eba611f65604a8