Dynamically register lint issues as inspections
Normally lint issues are registered in the IDE as inspections (which
requires registering them in android-plugin.xml as a
<globalInspection>, along with a custom class extending
AndroidLintInspectionBase.

This is still necessary for any issues that provide quickfixes, and is
still done for all the builtin lint issues since it's more efficient,
but this CL also adds fallback handling to any unknown issues
encountered on the fly: these are registered with the inspection
machinery using a wrapper class. This lets custom views get exactly
the same experience as builtin issues in the IDE: they show up in
options, they participate properly in the "More information" action
for error messages, and so on.

(This supercedes the older custom-view support, where there were two
builtin inspections, one for error-severity custom checks, and one for
warning-severity custom checks, with a generic explanation message and
a custom fix for showing the original check.)

This CL also switches the issue registry from a list to a set to make
the contains() check faster from isEnabled().

Test: Already covered by AndroidLintTest

Change-Id: Ia1c8d5c6b7d2905ef99308ad35e7c08a50cbbdb6