Added Resize target to LinearLayoutHandler
Added ResizeTarget for linear layout handler.

Also refromated some related files, wrote some documentation for the
Scene related methods and corrected warnings.

Corrected a typo in LayoutSceneBuilder: the for loop in clearChildTarget
was clearing the parents instead od the children (even if the target
might be cleared in the recursive call after when the child becomes the
parent)

Test: Ignored old tests that failed because. Will implement test based on
Target in a following CL.

Change-Id: I8bf2aadd82c5a75393ec265cfbc2bb7905057b1f