[AIA] Introduce support for provisioning
This CL adds a BeforeRunTask that makes the provisioning of the phone
before running an Instant App RunConfiguration.

All the code for provisioning is based on python scripts contained in
whsdk version 0.2.28 and should be changed if that logic is changed.

Provisioning is made by ProvisionRunner, which chacks if each one of the
apks present in the instant app sdk should be installed and tries to
install each one if needed.

All the apks that should be installed are implemented extending a base
class ProvisionPackage.

TODOs:
- Verify if it's necessary to clean app's data
- Define better when it's a AIA context
- Add logger
- Add UI to indicate that device is provisioning

Test: in a following CL ag/1968599

Change-Id: I65c4e17a201f5c46d9293194bf831c3608b1c719