Fix categorization of inspections
In 2017.1 inspection cateogries can be nested. Take advantage
of this to map lint categories (which are already nested) to
inspection categories. This removes all the ugly, redundant
   Android > Lint >
prefixes for all the inspection categories.

In the inspections window on the other hand there is no nesting,
so switch from a single "Android Lint" category to the top level
categories from lint, e.g. "Android Lint: Security", etc.

Test: Unit test included

Change-Id: I70fff85b2894fbc4b6ffb94fa13749c5bdd420e4