Add link to run-config dialog when perfa is disabled
This change exposes the validate() method in AndroidRunConfigurationBase,
which allows the run configuration UI to invoke  and bring to focus the
area of interest based on the validation results. The profiler UI relies
on this mechanism to navigate to the profiling tab.

TODO: include link to official doc in AndroidProfilersPanel

Bug: 35840177
Test: N/A
Change-Id: I2eb59af843836eee9407fe2609fc069e437fab4d