Dynamically generate android.os.Build
We will now copy android.os._Original_Build into android.os.Build. This
way, we avoid having android.os.Build in the classpath of the plugins.
Some libraries used by the plugins would rely on checking it
android.os.Build is present in the classpath to decide if they are
running Android.

Bug: http://b.android.com/210345
Test: Added new test for the LayoutlibClassLoader
Change-Id: I27ff23b5ec1cafe5d7141e13ef93ce587d1a2435