Fix broken source Attachments test in ant tests
A new version of guava was added to the offline repo that
does not contain source attachments causing this test to
fail. In Bazel the dependencies have been updated to use
a specified version of guava, but this does not work for
the ant tests. This CL pins the version of guava used to
one that does have sources.

Test: N/A
Change-Id: I5186db5d55dc52b219484906065b70c8e50f4ce3