Fix CoordinatorLayout scroll handling
Before this CL, CoordinatorLayout wouldn't handle scrolling at all and
the only way to check the nested scroll would be going to the included
layout. This is the default in the scrolling activity template.
Also, NestedScrollView wouldn't handle correctly scrolling back when
included in a CoordinatorLayout.

This unifies a bit the scroll handling and fixes the issue above.

Test: Modified tests for scrolling views. Manually tested
CoordinatorLayout

Change-Id: I0abd31b5033d6c1998e6a0e83cb541637f1a5a6c