Fix TemplateTest in the presence of a more recent plugin
The Template class performs a sync at the end of
rendering by default. This causes an issue if there
is a more recent plugin visible as it attempts to
trigger the upgrade UI which fails in headless mode.

Disabling the automatic sync is also consistent with the
aims of the test which allows the user to configure
whether they want a sync to occur or not via the
CHECK_LINT field - by default the test just builds the
project using gradle directly.

Fixes: http://b.android.com/240097

Test: N/A
Change-Id: Idbf12c2f5cbb10700df357e07746ae9052f49515