Enable TemplateTest in Bazel
The tests were failing under Bazel due to the use of
getCanonicalFile which broke the Bazel sandbox,
and the generation of too long filenames due to the
location of the Bazel sandbox. This CL fixes these by
removing an unnecessary prefix to the project name and
the call to getCanonicalFile.

It also adds a list of skippable parameters that can
safely be ignored when not running in comprehensive
mode, and only tests variable combinations when in
comprehensive mode to keep the time down on the PSQ.

Fixes: http://b.android.com/240097

Test: N/A
Change-Id: I70774a257431d175942f48c4eb581bf94122f101