[AVD Mgr] Reduce the default number of cores used by an AVD
When creating an AVD, the user can select the number of cores
to be used when that virtual device runs. The old way was to
recommend half the number of cores available on the host.

The new way limits the recommendation to 4 cores. The user can
still select a different number, but there will be little
benefit from doing so.

Test: Just manual

Change-Id: I7da9c7d18802549ff72236138b74340f3254d385