Rescope Scene to DesignSurface
Previously, we had a single DesignSurface and NlModel, and all of
SceneView, Display, Scene, and SceneManager, were duplicated (for the
design view and the blueprint view).
However, the Scene and SceneManager are always identical between the
two. In this change, the Scene and SceneManager are moved from the
(duplicated) SceneView to the (single) DesignSurface.
Notably, this enables the future moving of rendering logic into
SceneManager (originally attempted in
I3c19771818b25fdf63125e446f547544aeebb74b).

Test: updated appropriate tests to deal with refactor.

Change-Id: I19e9a09211436cfab13099a769550abbd4211e9c