Make UI test more resilient
This test generates a set of errors and compares them
against an expected output. However the order the errors
appear in does not seem to be deterministic and some
errors only appear during local development. This CL
changes the comparison to split the output and check that
it only contains a subset of the expected error messages.

Test: N/A
Change-Id: Iedd7bfe0f7af051dc60e87ea106d888a355803a0