Add support to multiple devices when using Legacy Profiler
LegacyProfilingHandler should handle multiple devices when profiling. We
were not doing that and that was causing Studio to freeze due to a
deadlock: we were waiting for a countdown from a latch, which should
happen in a callback of LegacyProfilingHandler. Sometimes, however, we
were calling the callback in a different LegacyProfilingHandler and,
consequently, not calling the latch's countdown ever.

Bug: 36360708
Test: Manual testing. I tested by running the failing scenarios again
      and observing they don't fail anymore. Automated tests for
      StudioLegacyCpuTraceProfiler are not there yet.

Change-Id: I6b4cdf857f85bb4b9919a6de0a996ad40656a52b