[APK Debugging] Added setup issues that are preserved between IDE sessions.
This CL introduces SetupIssue which stored in ApkFacet, allowing
setup issues to be stored in disk and preserved between IDE sessions.

For example, if the IDE throws an error because an APK needs an
Android platform that is not installed in the SDK Manager, that
error will be displayed to the user even when the user reopens the
project.

Test: JUnit test.
Change-Id: If2e1a6ffbde1f013e99033e685b568d6cdac4f3c