Fix IdeFrameFixture.requestFocusIfLost
If the focus is in the process of being lost by a closing dialog,
calling IdeFrameFixture.requestFocusIfLost would sometimes be ignored
because the focus was still held by a non-showing component of the
dialog when called. We check that the focus owner is a showing component
before returning true in requestFocusIfLost.

Test: HardwareProfileTest
Change-Id: I693d19261fd8c5089dae45c2a91ff2c5e2f37203