Add a Downloadable Font Service.
In the property inspector the android:fontFamily property now will
have an option for selecting a font from a dialog by selecting the
"More Fonts..." option in the bottom of the dropdown.

The dialog has 3 sections: project fonts, system fonts, and
downloadable fonts. Selecting a downloadable font will create a new
font resource in the res/font folder of the project.

The dialog can preview all the fonts provided by the Google fonts
service. A download service will download a directory (there is a
builtin fallback version). Then start downloading menu fonts which
are special ttf files than only holds glyphs to display the name of
the font. Whenever the user selects a font a preview is shown to the
right with the different versions of the file (misc weights).

This feature is currently hidden behind a debug version only flag.
It can be toggled on the wheel menu of the property inspector/table
in the layout designer.

Test: Tests will be added in a followup CL.

Change-Id: I495cb757f130c3079c087efb7ef4e2f3ce4b1429