Fix EspressoRecorderTest.addDependencyOnFly
The test failed because the instrumentation test could not be ready
when it was launched from "Run" -> "Run..." menu. Use a wait-until
to make sure it is ready before the test goes to the next step.

Keep and monitor the test in QA_unreliable for 20 runs before moving
to QA group.

Test: run the test.
Change-Id: I736915c6d8f31fb33a72eae9ebe3af737457298b