Add show bytecode action for dex classes and methods
This CL adds the ability to view the dex byte code for classes and methods.
This action can be invoked by right clicking on a class or method node when
viewing a dex file tree inside the APK Analyzer.  The implementation is just
a thin wrapper on top of baksmali which does the disassembly, and the
resulting byte code is displayed inside a popup that has an editor with the
file type set to smali.

Bug: b/35986213
Test: unit tests added

Change-Id: I9e99ca81f2eded9ff00670c3112389f6010e2fbe