Update packages used in project output
This CL updates the package names used to create modules
in a new project. The application package name is used for
the application module and the instant app module / splits
and new packages are used for the two libraries that are
created.

The new package naming is enabled for both the new and old
wizards and fixes a bug with module naming that was
preventing the new New Module Wizard flow from working.

This CL requires the new templates in the other CL in this
topic to function correctly

Fixes: http://b/36443685

Test: existing tests updated to new package scheme
Change-Id: I82324e54d83e6481bc8f012cca0fd94e6f47e66c