[AIA] Update provision based on last AIA SDK
This CL makes the necessary modifications in the ProvisionPackage s
based on last changes in AIA SDK. Based on whsdk-0.2.33

Policy packages have changed. Now, instead of installing a specific
policy set package, we check all the apks in the specific folder and
install them if needed.

This CL also changes the version checking of the packages from comparing
the versionName to comparing the versionCode.

Test: JUnit tests
Change-Id: I88a3565de294ec1685652835af8f81d7ab409ce9