Fixing SDK background installs
I99ca66a7100cbbc827276d6a9a4ccfe46b477b53 broke sdk component background
installs, in that the backgrounded install didn't know it was
backgrounded.
This is because the install task is a backgroundable task, but always
runs in the background and thus sentToBackground is never called (when
we supposedly "background" the task we actually just cancel the wizard
and let the task continue on).
Fixing by just making the wizard tell the task explicitly that it's
backgrounded.

Test: InstallTaskTest.testBackground
Change-Id: Ib582358b67820674478176628451e968bdcbf003