264038: Disable streams-iteration suggestion unless minSdkVersion allows it
The various Java 8 language feature inspections have been unregistered
on the IDE side. This CL registers subclasses of them insead, which
first check the API level and only if adequate continues on into the
original inspection. This ensures that we don't propose changes (such
as replacing Collections.sort(collection, comparator) with List.sort(
comparator) which requires API 24, unless minSdkVersion is at least 24.

Also includes language level 1.8 in the PSD panel unconditionally
(it was tried to the presence of the N preview SDK), and removes
some unused code.

Test: This CL includes a unit test which spot checks a couple of these
inspections.

Change-Id: Ie4e4f25d24d773871a7db6ce91a451359a27091f