Refactor DragTarget and Notch interaction
Renamed DragTarget to ConstraintDragTarget and made the notch
funcionnalities available to other Targets by introducing the
TargetNotchConnector.

The Notch.Action should also be transfered to Target to ensure that
Notches only deal with UI interaction.

Test: ran ConstraintLayoutTest but it was already marked as unrelaible
and didn't pass.
UI test for LinearLayoutHandler will be written and will cover this
refactoring

Change-Id: I068ecab74e8c9408308626ec16d3852033a6889b