[AIA] Introduce support for deploying in post O
The mechanism to run an instant app in a pre O device is different from
running in a post O (O and later) device: e.g. instead of using the
supervisor in post O, we install apps with --ephemeral flag.

This CL instroduces support for deploying in post O by checking what is
the android version in the device and making the necessary modifications
in the flow, running adb install with --ephemeral flag when needed.

Test: JUnit tests

Fixes bug http://b/35725963

Change-Id: Ia96cab19bd843a98113e9a098860464cdad47967