New Project Wizard: File > New Module - More than one option selected
We display a gallery chooser and when pressing next, only the step for
the item selected should be selected. This seems a regression
introduced by Change-Id: If35776f71ac4d2be477dafffa7e6b08ba725b8e2
When onProceeding() is called, a step is allowed to change the
visibility of its "children" steps, that means the next step cannot be
determined in advance.

Test: NewModuleTest@createNewModuleFromJar
Change-Id: Iaa0e87f00a110ac120785347fc70c55c8862fbae