[AIA] Update provision to work with SDK 0.2.35
The differences in the provision script are basically a new phone
architecture (armeabi-v7a) supported a shell command run to notify that
the content filters have changed.

This CL also changes the old SDK in testData to the new one. Since we
still don't have AIA SDK in the public SDK, we need it in testData for
the moment. Bug b/37083525 is tracking it to be deleted once the sdk
becomes public next month.

Test: current tests should pass with the new SDK (in testData)
Change-Id: I4121f035d82870d2b236c8162792af3bbd69597e