DOM unit tests should run without lint inspections enabled
This CL turns off lint inspections when other highlighting
tests are performed. The API used to be geared towards
enabling exactly one inspection; it's now changed to
disable all but a given set of inspections, so to make it
clearer I renamed it and made it take a varargs parameter.

Test: This CL only touches tests.
Change-Id: I5dac691cdd2464a5839f48e0cd2c49c3463ed88c