Removed custom link colors for Instant Run notification
In the "Edit Configuration..." dialog, under the General tab,
if you had an older version of the Android Gradle plugin, you
would see a message

"Instan Run requires a newer version of the Gradle plugin.
Update Project"

and the update project would be red with a blue underline.

However, there are two issues here:
1) The red style looks like an error more than info
2) There are other places in our tool that show this same
   notification (as other features also are introduced which
   depend on certain versions of the Gradle plugin) which
   don't know to use this custom look and feel. This makes our
   product feel inconsistent.

Test: N/A, UI change only
Change-Id: I043d79057d89e03cd1510b015900f3e93aecf0c7