[AIA] Consume ProjectBuildOutput from the IDE
When migrating to plugin 2.5, outputs are not known at sync time
anymore. To get the outputs and be able to install apks in the device,
we get the ProjectBuildOutput of a certain module right after the build
is performed when running a RunConfiguration.

This output model is used in GradleApkProvider to get the right apk
files.

Test: JUnit tests
Change-Id: I3c30facbc2509a7a3bfc0809c81b1bd75236b88e